dev to demo

+++ DOMAIN/Entities/Materials/Batch/MaterialBatchDto.cs
@@ -12,9 +12,9 @@ public class MaterialBatchDto
     public BatchStatus Status { get; set; }  
     public DateTime DateReceived { get; set; }
     public DateTime? DateApproved { get; set; }
-    public int TotalQuantity { get; set; }        
-    public int ConsumedQuantity { get; set; }  
-    public int RemainingQuantity { get; set; }
+    public decimal TotalQuantity { get; set; }        
+    public decimal ConsumedQuantity { get; set; }  
+    public decimal RemainingQuantity { get; set; }
     public DateTime ExpiryDate { get; set; }
     public List<MaterialBatchEventDto> Events { get; set; } = [];
     public List<MaterialBatchMovementDto> Movements { get; set; } = [];
@@ -24,7 +24,7 @@ public class MaterialBatchDto
 public class MaterialBatchEventDto 
 {
     public EventType Type { get; set; }
-    public int Quantity { get; set; }     
+    public decimal Quantity { get; set; }     
     public CollectionItemDto User { get; set; }          
     public DateTime CreatedAt { get; set; }    
 }
@@ -34,7 +34,7 @@ public class MaterialBatchMovementDto : BaseDto
     public CollectionItemDto Batch { get; set; }
     public CollectionItemDto FromLocation { get; set; }
     public CollectionItemDto ToLocation { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public DateTime MovedAt { get; set; }
     public CollectionItemDto MovedBy { get; set; }
     public MovementType MovementType { get; set; }  
@@ -43,5 +43,5 @@ public class MaterialBatchMovementDto : BaseDto
 public class CurrentLocationDto
 {
     public string LocationName { get; set; }
-    public int QuantityAtLocation { get; set; }
+    public decimal QuantityAtLocation { get; set; }
 }
