+++ DOMAIN/Entities/ProductionSchedules/CreateProductionScheduleRequest.cs
@@ -10,7 +10,7 @@ public class CreateProductionScheduleRequest
     public DateTime ScheduledEndTime { get; set; }
     public Guid? ProductId { get; set; }
     public ProductionStatus Status { get; set; } 
-    public int Quantity { get; set; } // Quantity of the product to be produced
+    public decimal Quantity { get; set; } // Quantity of the product to be produced
     [StringLength(1000)] public string Remarks { get; set; } // Optional remarks for additional notes
     public List<CreateProductionScheduleItemRequest> Items { get; set; } = [];
 }
@@ -19,5 +19,5 @@ public class CreateProductionScheduleItemRequest
 {
     public Guid MaterialId { get; set; }
     public Guid? UomId { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
 }
