+++ DOMAIN/Entities/Products/CreateProductBillOfMaterialRequest.cs
@@ -4,7 +4,7 @@ public class CreateProductBillOfMaterialRequest
 {
     public Guid? ComponentMaterialId { get; set; }
     public Guid? ComponentProductId { get; set; }
-    public int Quantity { get; set; }  // Quantity of the component required
+    public decimal Quantity { get; set; }  // Quantity of the component required
     public Guid UoMId { get; set; }  // Unit of Measure, e.g., grams, liters, pieces
     public bool IsSubstitutable { get; set; }  // Allows for substitution in production
 }
