+++ DOMAIN/Entities/WorkOrders/CreateWorkOrderRequest.cs
@@ -9,7 +9,7 @@ public class CreateWorkOrderRequest
     public Guid ProductId { get; set; }
     public Guid? MasterProductionScheduleId { get; set; } // Link to the Master Production Schedule
     
-    public int Quantity { get; set; } // Quantity of the product to be produced
+    public decimal Quantity { get; set; } // Quantity of the product to be produced
 
     public DateTime StartDate { get; set; } // Scheduled start date
     public DateTime EndDate { get; set; } // Scheduled end date
